requirements.txt
google-genai==1.38.0
openai>=1.40.0
